[Terry N.] don't clobber .xml and .md files in ./pduLogs during tests

+++ test/edu/nps/moves/dis7/test/AllPduRoundTripTest.java
@@ -56,7 +56,9 @@ public class AllPduRoundTripTest
     File file = new File("./pduLog");
 
     for (File subFile : file.listFiles()) {
-        subFile.delete();
+        
+        if (!subFile.getName().contains(".xml") && !subFile.getName().contains(".md"))
+            subFile.delete();
     }
   }
